added traits for broker
